+++ identity_credential/Cargo.toml
@@ -21,7 +21,6 @@ identity_document = { version = "=1.1.1", path = "../identity_document", default
 identity_verification = { version = "=1.1.1", path = "../identity_verification", default-features = false }
 indexmap = { version = "2.0", default-features = false, features = ["std", "serde"] }
 itertools = { version = "0.11", default-features = false, features = ["use_std"], optional = true }
-json-proof-token.workspace = true
 once_cell = { version = "1.18", default-features = false, features = ["std"] }
 reqwest = { version = "0.11", default-features = false, features = ["default-tls", "json", "stream"], optional = true }
 roaring = { version = "0.10.2", default-features = false, features = ["serde"], optional = true }
@@ -33,7 +32,8 @@ serde_repr = { version = "0.1", default-features = false, optional = true }
 strum.workspace = true
 thiserror.workspace = true
 url = { version = "2.5", default-features = false }
-zkryptium.workspace = true
+zkryptium = { workspace = true, optional = true }
+json-proof-token = { workspace = true, optional = true }
 
 [dev-dependencies]
 anyhow = "1.0.62"
@@ -57,4 +57,5 @@ status-list-2021 = ["revocation-bitmap"]
 validator = ["dep:itertools", "dep:serde_repr", "credential", "presentation"]
 domain-linkage = ["validator"]
 domain-linkage-fetch = ["domain-linkage", "dep:reqwest", "dep:futures"]
-sd-jwt = ["credential", "validator", "sd-jwt-payload"]
+sd-jwt = ["credential", "validator", "dep:sd-jwt-payload"]
+jpt-bbs-plus = ["credential", "validator", "dep:zkryptium", "dep:json-proof-token"]
