+++ identity_credential/src/revocation/mod.rs
@@ -9,9 +9,11 @@ mod revocation_bitmap_2022;
 #[cfg(feature = "status-list-2021")]
 pub mod status_list_2021;
 
+#[cfg(feature = "jpt-bbs-plus")]
 pub mod validity_timeframe_2024;
 
 pub use self::error::RevocationError;
 pub use self::error::RevocationResult;
 pub use revocation_bitmap_2022::*;
+#[cfg(feature = "jpt-bbs-plus")]
 pub use validity_timeframe_2024::*;
