+++ identity_iota/Cargo.toml
@@ -64,6 +64,9 @@ memstore = ["identity_storage/memstore"]
 # Enables selective disclosure features.
 sd-jwt = ["identity_credential/sd-jwt"]
 
+# Enables zero knowledge selective disclosurable VCs
+jpt-bbs-plus = ["identity_storage/jpt-bbs-plus", "identity_credential/jpt-bbs-plus"]
+
 [package.metadata.docs.rs]
 # To build locally:
 # RUSTDOCFLAGS="--cfg docsrs" cargo +nightly doc --all-features --no-deps --workspace --open
