+++ identity_storage/Cargo.toml
@@ -21,14 +21,14 @@ identity_document = { version = "=1.1.1", path = "../identity_document", default
 identity_iota_core = { version = "=1.1.1", path = "../identity_iota_core", default-features = false, optional = true }
 identity_verification = { version = "=1.1.1", path = "../identity_verification", default_features = false }
 iota-crypto = { version = "0.23", default-features = false, features = ["ed25519"], optional = true }
-json-proof-token.workspace = true
 rand = { version = "0.8.5", default-features = false, features = ["std", "std_rng"], optional = true }
 seahash = { version = "4.1.0", default_features = false }
 serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
 tokio = { version = "1.29.0", default-features = false, features = ["macros", "sync"], optional = true }
-zkryptium.workspace = true
+zkryptium = { workspace = true, optional = true }
+json-proof-token = { workspace = true, optional = true }
 
 [dev-dependencies]
 identity_credential = { version = "=1.1.1", path = "../identity_credential", features = ["revocation-bitmap"] }
@@ -44,3 +44,5 @@ memstore = ["dep:tokio", "dep:rand", "dep:iota-crypto"]
 send-sync-storage = []
 # Implements the JwkStorageDocumentExt trait for IotaDocument
 iota-document = ["dep:identity_iota_core"]
+# Enables JSON Proof Token & BBS+ related features
+jpt-bbs-plus = ["identity_credential/jpt-bbs-plus", "dep:zkryptium", "dep:json-proof-token"]
