+++ identity_storage/src/key_storage/jwk_storage.rs
@@ -4,12 +4,9 @@
 use crate::key_storage::KeyId;
 use crate::key_storage::KeyStorageError;
 use crate::key_storage::KeyType;
-use crate::ProofUpdateCtx;
 use async_trait::async_trait;
 use identity_verification::jose::jwk::Jwk;
 use identity_verification::jose::jws::JwsAlgorithm;
-use jsonprooftoken::jpa::algs::ProofAlgorithm;
-use zkryptium::bbsplus::signature::BBSplusSignature;
 
 use super::jwk_gen_output::JwkGenOutput;
 
@@ -64,31 +61,4 @@ pub trait JwkStorage: storage_sub_trait::StorageSendSyncMaybe {
 
   /// Returns `true` if the key with the given `key_id` exists in storage, `false` otherwise.
   async fn exists(&self, key_id: &KeyId) -> KeyStorageResult<bool>;
-}
-
-/// Extension to the JwkStorage to handle BBS+ keys
-#[cfg_attr(not(feature = "send-sync-storage"), async_trait(?Send))]
-#[cfg_attr(feature = "send-sync-storage", async_trait)]
-pub trait JwkStorageExt: JwkStorage {
-  /// Generates a JWK representing a BBS+ signature
-  async fn generate_bbs(&self, key_type: KeyType, alg: ProofAlgorithm) -> KeyStorageResult<JwkGenOutput>;
-
-  /// Sign the provided `data` and `header` using the private key identified by `key_id` according to the requirements
-  /// of the corresponding `public_key` (see [`Jwk::alg`](Jwk::alg()) etc.).
-  async fn sign_bbs(
-    &self,
-    key_id: &KeyId,
-    data: &[Vec<u8>],
-    header: &[u8],
-    public_key: &Jwk,
-  ) -> KeyStorageResult<Vec<u8>>;
-
-  /// Update proof functionality for timeframe revocation mechanism
-  async fn update_signature(
-    &self,
-    key_id: &KeyId,
-    public_key: &Jwk,
-    signature: &[u8; BBSplusSignature::BYTES],
-    ctx: ProofUpdateCtx,
-  ) -> KeyStorageResult<[u8; BBSplusSignature::BYTES]>;
 }
