+++ identity_storage/src/key_storage/mod.rs
@@ -6,12 +6,14 @@
 //! This module provides the [`JwkStorage`] trait that
 //! abstracts over storages that store JSON Web Keys.
 
-#[cfg(feature = "memstore")]
+#[cfg(all(feature = "memstore", feature = "jpt-bbs-plus"))]
 mod bls;
 #[cfg(feature = "memstore")]
 mod ed25519;
 mod jwk_gen_output;
 mod jwk_storage;
+#[cfg(feature = "jpt-bbs-plus")]
+mod jwk_storage_bbs_plus_ext;
 mod key_id;
 mod key_storage_error;
 mod key_type;
@@ -23,6 +25,8 @@ pub(crate) mod tests;
 
 pub use jwk_gen_output::*;
 pub use jwk_storage::*;
+#[cfg(feature = "jpt-bbs-plus")]
+pub use jwk_storage_bbs_plus_ext::*;
 pub use key_id::*;
 pub use key_storage_error::*;
 pub use key_type::*;
