+++ identity_stronghold/Cargo.toml
@@ -21,9 +21,9 @@ iota_stronghold = { version = "2.0.0", default-features = false }
 rand = { version = "0.8.5", default-features = false, features = ["std", "std_rng"] }
 tokio = { version = "1.29.0", default-features = false, features = ["macros", "sync"] }
 zeroize = { version = "1.6.0", default_features = false }
-zkryptium.workspace = true
-json-proof-token.workspace = true
 anyhow = "1.0.82"
+zkryptium = { workspace = true, optional = true }
+json-proof-token = { workspace = true, optional = true }
 
 
 [dev-dependencies]
@@ -34,3 +34,4 @@ tokio = { version = "1.29.0", default-features = false, features = ["macros", "s
 default = []
 # Enables `Send` + `Sync` bounds for the trait implementations on `StrongholdStorage`.
 send-sync-storage = ["identity_storage/send-sync-storage"]
+bbs-plus = ["identity_storage/jpt-bbs-plus", "dep:zkryptium", "dep:json-proof-token"]
