+++ identity_stronghold/src/lib.rs
@@ -3,6 +3,7 @@
 
 pub(crate) mod ed25519;
 mod stronghold_jwk_storage;
+#[cfg(feature = "bbs-plus")]
 mod stronghold_jwk_storage_ext;
 mod stronghold_key_id;
 pub(crate) mod stronghold_key_type;
